+++ ajax/data_cluster.php
@@ -25,7 +25,8 @@ $jumlahData = $resultTotal[0]['jumlah'];
 $jumlahHalaman = ceil($jumlahData / $jumlahDataPerHalaman);
 
 // Mendefinisikan tautan pagination secara langsung
-$pagination = '<ul class="pagination justify-content-end">';
+$pagination = '<span id="showing-entries">Showing ' . ($awalData + 1) . ' to ' . min($awalData + $jumlahDataPerHalaman, $jumlahData) . ' of ' . $jumlahData . ' entries</span>';
+$pagination .= '<ul class="pagination pagination-sm m-0 justify-content-end">';
 $pagination .= '<li class="page-item"><a class="page-link" href="?page=' . max(1, $page - 1) . '">Previous</a></li>';
 
 $jumlahTampil = min(5, $jumlahHalaman);
@@ -39,10 +40,9 @@ for ($i = $start; $i <= $end; $i++) {
         $pagination .= '<li class="page-item"><a class="page-link" href="?page=' . $i . '">' . $i . '</a></li>';
     }
 }
-
-
 $pagination .= '<li class="page-item"><a class="page-link" href="?page=' . min($jumlahHalaman, $page + 1) . '">Next</a></li>';
 $pagination .= '</ul>';
+$pagination .= '</div>';
 ?>
 
 <table class="table color-table info-table">
@@ -53,25 +53,31 @@ $pagination .= '</ul>';
             <th></th>
         </tr>
     </thead>
-    <?php foreach ($cluster as $row) : ?>
-        <tbody>
-            <tr>
-                <td><?= $row["id_cluster"]; ?></td>
-                <td><?= $row["nama_cluster"]; ?></td>
-                <td>
-                    <div class="dropdown">
-                        <button class="btn dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
-                            Action
-                        </button>
-                        <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">
-                            <li><a class="dropdown-item" href="edit_cluster.php?id_cluster=<?= $row["id_cluster"]; ?>">Edit</a></li>
-                            <li><a class="dropdown-item tombol-hapus" href="delete_cluster.php?id_cluster=<?= $row["id_cluster"]; ?>">Delete</a></li>
-                        </ul>
-                    </div>
-                </td>
-            </tr>
-        </tbody>
-    <?php endforeach; ?>
+    <?php if ($jumlahData > 0): ?>
+        <?php foreach ($cluster as $row) : ?>
+            <tbody>
+                <tr>
+                    <td><?= $row["id_cluster"]; ?></td>
+                    <td><?= $row["nama_cluster"]; ?></td>
+                    <td>
+                        <div class="dropdown">
+                            <button class="btn dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
+                                Action
+                            </button>
+                            <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">
+                                <li><a class="dropdown-item" href="edit_cluster.php?id_cluster=<?= $row["id_cluster"]; ?>">Edit</a></li>
+                                <li><a class="dropdown-item tombol-hapus" href="delete_cluster.php?id_cluster=<?= $row["id_cluster"]; ?>">Delete</a></li>
+                            </ul>
+                        </div>
+                    </td>
+                </tr>
+            </tbody>
+        <?php endforeach; ?>
+    <?php else: ?>
+        <tr>
+            <td colspan="9" class="text-center">No data found</td>
+        </tr>
+    <?php endif; ?>
 </table>
 <!-- Tampilkan pagination -->
 <?= $pagination ?>
