+++ ajax/data_users.php
@@ -32,7 +32,8 @@ $jumlahData = $resultTotal[0]['jumlah'];
 $jumlahHalaman = ceil($jumlahData / $jumlahDataPerHalaman);
 
 // Mendefinisikan tautan pagination secara langsung
-$pagination = '<ul class="pagination justify-content-end">';
+$pagination = '<span id="showing-entries">Showing ' . ($awalData + 1) . ' to ' . min($awalData + $jumlahDataPerHalaman, $jumlahData) . ' of ' . $jumlahData . ' entries</span>';
+$pagination .= '<ul class="pagination pagination-sm m-0 justify-content-end">';
 $pagination .= '<li class="page-item"><a class="page-link" href="?page=' . max(1, $page - 1) . '">Previous</a></li>';
 
 $jumlahTampil = min(5, $jumlahHalaman);
@@ -48,6 +49,7 @@ for ($i = $start; $i <= $end; $i++) {
 }
 $pagination .= '<li class="page-item"><a class="page-link" href="?page=' . min($jumlahHalaman, $page + 1) . '">Next</a></li>';
 $pagination .= '</ul>';
+$pagination .= '</div>';
 
 ?>
 
@@ -66,36 +68,42 @@ $pagination .= '</ul>';
         </thead>
         <tbody>
             <?php $n = 1; ?>
-            <?php foreach ($users as $userData) : ?>
+            <?php if ($jumlahData > 0): ?>
+                <?php foreach ($users as $userData) : ?>
+                    <tr>
+                        <td><?= $n; ?></td>
+                        <td><?= $userData["id"]; ?></td>
+                        <td><?= $userData["username"]; ?></td>
+                        <td><?= $userData["nama"]; ?></td>
+                        <td><?= $userData["email"]; ?></td>
+                        <td>
+                            <?php
+                            if ($userData['role'] == 'Admin') {
+                                echo '<div class="label label-table label-success">' . $userData["role"] . '</div>';
+                            } else {
+                                echo '<div class="label label-table label-info">' . $userData["role"] . '</div>';
+                            }
+                            ?>
+                        </td>
+                        <td>
+                            <div class="dropdown">
+                                <button class="btn dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
+                                    Action
+                                </button>
+                                <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">
+                                    <li><a class="dropdown-item" href="edit_users.php?id=<?= $userData["id"]; ?>">Edit</a></li>
+                                    <li><a class="dropdown-item tombol-hapus" href="delete_users.php?id=<?= $userData["id"]; ?>">Delete</a></li>
+                                </ul>
+                            </div>
+                        </td>
+                    </tr>
+                    <?php $n++; ?>
+                <?php endforeach; ?>
+            <?php else: ?>
                 <tr>
-                    <td><?= $n; ?></td>
-                    <td><?= $userData["id"]; ?></td>
-                    <td><?= $userData["username"]; ?></td>
-                    <td><?= $userData["nama"]; ?></td>
-                    <td><?= $userData["email"]; ?></td>
-                    <td>
-                        <?php
-                        if ($userData['role'] == 'Admin') {
-                            echo '<div class="label label-table label-success">' . $userData["role"] . '</div>';
-                        } else {
-                            echo '<div class="label label-table label-info">' . $userData["role"] . '</div>';
-                        }
-                        ?>
-                    </td>
-                    <td>
-                        <div class="dropdown">
-                            <button class="btn dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
-                                Action
-                            </button>
-                            <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">
-                                <li><a class="dropdown-item" href="edit_users.php?id=<?= $userData["id"]; ?>">Edit</a></li>
-                                <li><a class="dropdown-item tombol-hapus" href="delete_users.php?id=<?= $userData["id"]; ?>">Delete</a></li>
-                            </ul>
-                        </div>
-                    </td>
+                    <td colspan="9" class="text-center">No data found</td>
                 </tr>
-                <?php $n++; ?>
-            <?php endforeach; ?>
+            <?php endif; ?>
         </tbody>
     </table>
 </div>
